+++ bin/nanocl/src/commands/cargo_image.rs
@@ -12,15 +12,28 @@ use crate::utils;
 use crate::models::{
   CargoImageOpts, CargoImageCommands, CargoImageRemoveOpts,
   CargoImageInspectOpts, CargoImageRow, CargoImageImportOpts,
+  CargoImageListOpts,
 };
 
-async fn exec_cargo_image_ls(client: &NanocldClient) -> IoResult<()> {
-  let items = client.list_cargo_image(None).await?;
+async fn exec_cargo_image_ls(
+  opts: &CargoImageListOpts,
+  client: &NanocldClient,
+) -> IoResult<()> {
+  let items = client.list_cargo_image(Some(opts.clone().into())).await?;
   let rows = items
     .into_iter()
     .map(CargoImageRow::from)
     .collect::<Vec<CargoImageRow>>();
-  utils::print::print_table(rows);
+  match opts.quiet {
+    true => {
+      for row in rows {
+        println!("{}", row.id);
+      }
+    }
+    false => {
+      utils::print::print_table(rows);
+    }
+  }
   Ok(())
 }
 
@@ -75,7 +88,7 @@ fn update_progress(
   }
 }
 
-pub(crate) async fn exec_cargo_image_create(
+pub(crate) async fn exec_cargo_image_pull(
   client: &NanocldClient,
   name: &str,
 ) -> IoResult<()> {
@@ -150,12 +163,12 @@ pub async fn exec_cargo_image(
   cmd: &CargoImageOpts,
 ) -> IoResult<()> {
   match &cmd.commands {
-    CargoImageCommands::List => exec_cargo_image_ls(client).await,
+    CargoImageCommands::List(opts) => exec_cargo_image_ls(opts, client).await,
     CargoImageCommands::Inspect(opts) => {
       exec_cargo_image_inspect(client, opts).await
     }
-    CargoImageCommands::Create(opts) => {
-      exec_cargo_image_create(client, &opts.name).await
+    CargoImageCommands::Pull(opts) => {
+      exec_cargo_image_pull(client, &opts.name).await
     }
     CargoImageCommands::Remove(args) => exec_cargo_image_rm(client, args).await,
     CargoImageCommands::Import(opts) => {