+++ bin/nanocl/src/commands/system.rs
@@ -1,7 +1,11 @@
 use nanocl_utils::io_error::IoResult;
 use nanocld_client::NanocldClient;
 
-use crate::models::{ProcessOpts, ProcessRow};
+use crate::models::{
+  ProcessOpts, ProcessRow, SystemOpts, SystemHttpOpts, SystemHttpCommands,
+  SystemCommands,
+};
+use crate::utils;
 use crate::utils::print::print_table;
 
 pub async fn exec_process(
@@ -19,4 +23,27 @@ pub async fn exec_process(
   print_table(rows);
 
   Ok(())
+}
+
+pub async fn exec_http(
+  client: &NanocldClient,
+  opts: &SystemHttpOpts,
+) -> IoResult<()> {
+  match &opts.commands {
+    SystemHttpCommands::Logs(opts) => {
+      let logs = client.list_http_metric(Some(opts.clone().into())).await?;
+      utils::print::print_yml(logs)?;
+    }
+  }
+
+  Ok(())
+}
+
+pub async fn exec_system(
+  client: &NanocldClient,
+  opts: &SystemOpts,
+) -> IoResult<()> {
+  match &opts.commands {
+    SystemCommands::Http(opts) => exec_http(client, opts).await,
+  }
 }