+++ bin/nanocl/src/models/state.rs
@@ -16,18 +16,31 @@ pub struct StateBuildArgs {
 }
 
 #[derive(Debug, Parser)]
-pub struct StateOpts {
-  /// Path or url to the state
-  #[clap(long, short = 'f')]
-  pub file_path: Option<String>,
-  /// Force pulling images
+pub struct StateApplyOpts {
+  /// Path or Url to the StateFile
+  #[clap(long, short = 's')]
+  pub state_location: Option<String>,
+  /// Force pulling images even if they exist
   #[clap(long, short = 'p')]
   pub force_pull: bool,
-  /// Attach to the logs of the deployed cargo when applying a state
-  #[clap(long, short = 'a')]
-  pub attach: bool,
+  /// Follow logs of the deployed cargo
+  #[clap(long, short = 'f')]
+  pub follow: bool,
   /// Skip the confirmation prompt
-  #[clap(short = 'y')]
+  #[clap(long = "yes", short = 'y')]
+  pub skip_confirm: bool,
+  /// Additional arguments to pass to the file
+  #[clap(last = true, raw = true)]
+  pub args: Vec<String>,
+}
+
+#[derive(Debug, Parser)]
+pub struct StateRemoveOpts {
+  /// Path or Url to the StateFile
+  #[clap(long, short = 's')]
+  pub state_location: Option<String>,
+  /// Skip the confirmation prompt
+  #[clap(long = "yes", short = 'y')]
   pub skip_confirm: bool,
   /// Additional arguments to pass to the file
   #[clap(last = true, raw = true)]
@@ -36,13 +49,14 @@ pub struct StateOpts {
 
 #[derive(Debug, Subcommand)]
 pub enum StateCommands {
-  /// Apply a state from a configuration file
-  Apply(StateOpts),
-  /// Revert a state from a configuration file
-  Revert(StateOpts),
+  /// Create or Update elements from a StateFile
+  Apply(StateApplyOpts),
+  /// Remove elements from a StateFile
+  #[clap(alias("rm"))]
+  Remove(StateRemoveOpts),
 }
 
-/// Manage configuration states
+/// Define, Run, or Remove Cargo or Virtual Machines
 #[derive(Debug, Parser)]
 pub struct StateArgs {
   #[clap(subcommand)]