+++ bin/nanocl/src/utils/mod.rs
@@ -1,7 +1,6 @@
 pub mod state;
 pub mod print;
 pub mod math;
-pub mod network;
 pub mod docker;
 pub mod installer;
 pub mod dialog;