+++ bin/nanocld/src/services/openapi.rs
@@ -48,13 +48,14 @@ use nanocl_stubs::vm_config::{
   VmConfig, VmConfigPartial, VmConfigUpdate, VmDiskConfig, VmHostConfig,
 };
 use nanocl_stubs::resource::{
-  Resource, ResourcePatch, ResourceConfig, ResourcePartial,
+  Resource, ResourceUpdate, ResourceConfig, ResourcePartial,
 };
 use nanocl_stubs::dns::{ResourceDnsRule, DnsEntry};
 use nanocl_stubs::proxy::{
   ResourceProxyRule, ProxyRuleHttp, ProxyHttpLocation, ProxySslConfig,
   ProxyRuleStream, StreamTarget, ProxyStreamProtocol, UriTarget,
-  LocationTarget, HttpTarget, UrlRedirect, CargoTarget, ProxyRule, UnixTarget,
+  LocationTarget, HttpTarget, UrlRedirect, UpstreamTarget, ProxyRule,
+  UnixTarget,
 };
 
 use super::{
@@ -259,7 +260,7 @@ impl Modify for VersionModifier {
     resource::inspect_resource,
     resource::create_resource,
     resource::delete_resource,
-    resource::patch_resource,
+    resource::put_resource,
     resource::list_resource_history,
     resource::revert_resource,
     // Metric
@@ -382,7 +383,7 @@ impl Modify for VersionModifier {
     VmHostConfig,
     // Resource
     Resource,
-    ResourcePatch,
+    ResourceUpdate,
     ResourceConfig,
     ResourcePartial,
     // ProxyRules
@@ -397,7 +398,7 @@ impl Modify for VersionModifier {
     LocationTarget,
     HttpTarget,
     UrlRedirect,
-    CargoTarget,
+    UpstreamTarget,
     UnixTarget,
     UriTarget,
     // DnsRules