+++ bin/ncdproxy/changelog.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [untagged]
 
+### Changed
+
+- Target use generic Target `Key`, `Port` and `Watch` require to specify if it's cargo or vm with `.c` or `.v`
+
+### Added
+
+- Options `DisableLogging` and `Path` for `CargoTarget` to disable logging for specific path.
+
 ### Fixed
 
 - Thread background crash if `/var/log/nginx/access` doesn't exists.
+- Updating a cargo wasn't refreshing the nginx config when using `nanocl cargo revert`.
 
 ## [0.4.3] - 2023-05-14